backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Rate Limiting and Security
slowapi==0.1.9

# CORS and Security Headers
fastapi-cors==0.0.6

# Environment and Configuration
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Upload and Storage
pillow==10.1.0

# Date and Time
python-dateutil==2.8.2

# Utilities
python-slugify==8.0.1

# 2FA Support
pyotp==2.9.0
qrcode[pil]==7.4.2
